magento/magento2#22447 integration-test-fix
