+++ dev/tests/integration/testsuite/Magento/Config/Model/Config/Structure/Reader/_files/catalog_new_products_list.xml
