+++ dev/tests/integration/testsuite/Magento/Config/Model/Config/Structure/Reader/_files/config.xml
