+++ dev/tests/integration/testsuite/Magento/Config/Model/Config/Structure/Reader/_files/orders_and_returns_customized.xml
